refactor: result페이지 앱라우터로 이동
