+++ src/app/generate/page.tsx
@@ -227,7 +227,7 @@ function GeneratePage() {
         const rid = await saveToSupabase(result);
         console.log('%c✅ Supabase 저장 성공:', 'color: green; font-weight: bold;');
         if (rid) {
-            router.push(`/result/result?rid=${encodeURIComponent(rid)}`);
+            router.push(`/result?rid=${encodeURIComponent(rid)}`);
         } else {
             console.error('❌ Request ID 생성 실패');
             alert('요청 ID 생성에 실패했습니다.');
